Example nightly bump commit
Ran `/automation/update-from-application-services.py nightly` and
checked in the results.  This is roughly what a nightly bump commit
would look like.

Note: the app-services side of this is basically just a mock.  It
doesn't actually build the xcframework or generate the swift code, it
just downloads them from the 97.2.0 release page.  Once we get a mac
worker we can test this for real.